+++ Assets/AppCenter/Plugins/AppCenterSDK/Distribute/Shared/Distribute.cs
@@ -79,6 +79,19 @@ namespace Microsoft.AppCenter.Unity.Distribute
             get; set;
         }
 
+        /// Sets the no release available callback.
+        /// </summary>
+        /// <value>The no release available callback.</value>
+        public static NoReleaseAvailableCallback NoReleaseAvailable
+        {
+            get; set;
+        }
+
+        public static void StartDistribute()
+        {
+            DistributeInternal.StartDistribute();
+        }
+
         /// <summary>
         /// If update dialog is customized by returning <c>true</c> in <see cref="ReleaseAvailableCallback"/>,
         /// You need to tell the distribute SDK using this function what is the user action.