+++ Assets/AppCenter/Plugins/AppCenterSDK/Distribute/Shared/ReleaseAvailableCallback.cs
@@ -9,7 +9,12 @@ namespace Microsoft.AppCenter.Unity.Distribute
     public delegate bool ReleaseAvailableCallback(ReleaseDetails releaseDetails);
 
     /// <summary>
+
     /// App will exit callback.
     /// </summary>
     public delegate void WillExitAppCallback();
+
+    /// No Release available callback.
+    /// </summary>
+    public delegate void NoReleaseAvailableCallback();
 }