+++ Assets/AppCenter/Plugins/AppCenterSDK/Distribute/iOS/DistributeInternal.cs
@@ -32,6 +32,15 @@ namespace Microsoft.AppCenter.Unity.Distribute.Internal
             return Distribute.ReleaseAvailable.Invoke(releaseDetails);
         }
 
+        delegate void NoReleaseAvailableDelegate();
+
+        static NoReleaseAvailableDelegate noReleaseDel;
+        [MonoPInvokeCallback(typeof(NoReleaseAvailableDelegate))]
+        static void NoReleaseAvailableFunc()
+        {
+            Distribute.NoReleaseAvailable?.Invoke();
+        }
+
         [MonoPInvokeCallback(typeof(WillExitAppDelegate))]
         static void WillExitAppFunc()
         {
@@ -55,8 +64,15 @@ namespace Microsoft.AppCenter.Unity.Distribute.Internal
             appcenter_unity_distribute_set_delegate();
             del = ReleaseAvailableFunc;
             willExitAppDel = WillExitAppFunc;
+            noReleaseDel = NoReleaseAvailableFunc;
             appcenter_unity_distribute_set_release_available_impl(del);
+            appcenter_unity_distribute_set_no_release_available_impl(noReleaseDel);
             appcenter_unity_distribute_set_will_exit_app_impl(willExitAppDel);
+        }
+
+        public static void StartDistribute()
+        {
+            appcenter_unity_start_distribute();
         }
 
         public static void AddNativeType(List<IntPtr> nativeTypes)
@@ -123,6 +139,9 @@ namespace Microsoft.AppCenter.Unity.Distribute.Internal
         private static extern void appcenter_unity_distribute_set_will_exit_app_impl(WillExitAppDelegate functionPtr);
 
         [DllImport("__Internal")]
+        private static extern void appcenter_unity_distribute_set_no_release_available_impl(NoReleaseAvailableDelegate functionPtr);
+
+        [DllImport("__Internal")]
         private static extern void appcenter_unity_distribute_replay_release_available();
 
         [DllImport("__Internal")]
@@ -130,6 +149,9 @@ namespace Microsoft.AppCenter.Unity.Distribute.Internal
 
         [DllImport("__Internal")]
         private static extern void appcenter_unity_distribute_check_for_update();
+
+        [DllImport("__Internal")]
+        private static extern void appcenter_unity_start_distribute();
 
 #endregion
     }