+++ Assets/AppCenter/Plugins/iOS/Distribute/DistributeDelegate.m
@@ -10,6 +10,7 @@
 @property NSObject *lockObject;
 @property ReleaseAvailableFunction releaseAvailableHandler;
 @property WillExitAppFunction willExitAppHandler;
+@property NoReleaseAvailableFunction noReleaseAvailableHandler;
 
 @end
 
@@ -44,6 +45,11 @@
   @synchronized (_lockObject) {
     _willExitAppHandler = implementation;
   }
+
+- (void)setNoReleaseAvailableImplementation:(NoReleaseAvailableFunction)implementation {
+  @synchronized (_lockObject) {
+    _noReleaseAvailableHandler = implementation;
+  }
 }
 
 - (BOOL)distribute:(MSACDistribute *)distribute releaseAvailableWithDetails:(MSACReleaseDetails *)details {
@@ -67,7 +73,7 @@
   }
 }
 
-- (void) replayReleaseAvailable {
+- (void)replayReleaseAvailable {
   ReleaseAvailableFunction handlerCopy = nil;
   MSACReleaseDetails *unprocessedCopy = _unprocessedDetails;
   @synchronized (_lockObject) {
@@ -76,6 +82,17 @@
   }
   if (unprocessedCopy && handlerCopy) {
     handlerCopy(unprocessedCopy);
+  }
+}
+
+- (void)distributeNoReleaseAvailable:(MSACDistribute *)distribute {
+  NoReleaseAvailableFunction handlerCopy = nil;
+  @synchronized (_lockObject) {
+    handlerCopy = _noReleaseAvailableHandler;
+    if (handlerCopy) {
+      handlerCopy();
+    }
+    return;
   }
 }