+++ Assets/AppCenter/Plugins/iOS/Distribute/DistributeUnity.mm
@@ -2,6 +2,7 @@
 // Licensed under the MIT license.
 
 #import "DistributeUnity.h"
+#import <AppCenter/MSACAppCenter.h>
 #import "DistributeDelegate.h"
 #import <Foundation/Foundation.h>
 
@@ -48,6 +49,16 @@ void appcenter_unity_distribute_set_release_available_impl(ReleaseAvailableFunct
 void appcenter_unity_distribute_set_will_exit_app_impl(WillExitAppFunction function)
 {
   [[UnityDistributeDelegate sharedInstance] setWillExitAppImplementation:function];
+}
+
+void appcenter_unity_distribute_set_no_release_available_impl(NoReleaseAvailableFunction function)
+{
+  [[UnityDistributeDelegate sharedInstance] setNoReleaseAvailableImplementation:function];
+}
+
+void appcenter_unity_start_distribute()
+{
+  [MSACAppCenter startService:MSACDistribute.class];
 }
 
 void appcenter_unity_distribute_replay_release_available()